+++ transparency-in-coverage/python/processors/utils/schema.py
